+++ client/cmdmain.h
@@ -18,6 +18,7 @@
 
 extern void UsbCommandReceived(UsbCommand *UC);
 extern int CommandReceived(char *Cmd);
+extern bool WaitForResponseTimeoutW(uint32_t cmd, UsbCommand* response, size_t ms_timeout, bool show_warning);
 extern bool WaitForResponseTimeout(uint32_t cmd, UsbCommand* response, size_t ms_timeout);
 extern bool WaitForResponse(uint32_t cmd, UsbCommand* response);
 extern void clearCommandBuffer();